requirements.txt
selenium>=4.16.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.2.0
lxml>=5.1.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
google-generativeai>=0.8.0
networkx>=3.2
plotly>=5.18.0
python-louvain>=0.16
